Refine helpers for the dynamic client.
This change creates a new wrapper for the dynamic client to allow
usage to mirror that of regular clients, e.g.

  client.Resource("mynamespace").Get(...)

This change also replaces NewUnstructuredInformer with
NewResourceInformer, removing unnecessary casting to Unstructured.
This will simplify integration of informers using the dynamic client
into the controller (which expects pkgruntime.Object).