Update TargetRef from name and kind to only TargetKind
The idea of the API intends to have a a set of preferences which can
be used for a target FederatedDeployment or FederatedReplicaset.
Although the set of preferences in question can be applied to multiple
target objects using label selectors, but there are no clear advantages
of doing that as of now.
To keep the implementation and usage simple, matching ns/name of RSP
resource to the target resource is sufficient and only additional information
needed in RSP resource is target kind (FederatedDeployment or FederatedReplicaset).