Bump e2e timeout to avoid deployment e2e timeout
Deployments have the most complex workflow of the existing kubernetes
types, and the interaction of the deployment controller with the sync
controller is likely to case the deployment crud test to take longer
than for other types.  Since the deployment crud test has been
regularly flaking, the previous timeout was likely not long enough.