Fix crudtester test of placement change
A namespace can't be removed from the host cluster via
placement since the namespace in the host cluster is effectively the
template.  Previously the crudtester's testing of a placement change
was applying this logic (don't remove the host cluster) to all types,
which prevented the placement change from being tested where the
deployment under test was a single cluster joined to itself.  This
commit ensures that logic applies only to namespaces, and localizes
the check of changing placement to simplify troubleshooting.