Fix crudtester version retrieval
Previously the crudtester was retrieving the PropagatedVersion for a
resource once and then checking that target resources matched.  In
cases where target resources are updated by a controller other than
the sync controller (e.g. deployments/jobs/replicasets), this created
a race between the crudtester and the other controller.  This commit
updates the crudtester to retrieve the propagated version on every
iteration of the wait loop to remove the possibility of a race
condition.