Refactor type registration for RSP
Ensure a mapping of type config name -> scheduling type since only a
single scheduler should be running for a given type.  Using the type
config name instead of the template kind ensures that scheduling will
work regardless of the naming of the federation primitives, since the
of a type config is always <plural name>.<group name> of the target.