Enable configurable replicas retention by the sync controller
`kubefed2 enable` is updated to add a `retainReplicas` field to any
type that exposes a `replicas` field.  If set on a federated resource,
the sync controller will attempt to retain a replicas field from the
target resource when performing updates to the target.  This is
intended to avoid thrashing the scheduler when in-cluster HPA is
targeting a scalable resource.

The granularity of this retention is all clusters or no-clusters, with
the expectation that where HPA is intended to be applied selectively,
a user will need to define 2 resources (one with `retainReplicas: true`
and one with `retainReplicas: false`) and define appropriate placement.