download-binaries.sh: Fix a problem with pipefail
If you are unlucky enough, "etcd --version | head -n 1" can
fail due to pipefail. (Depending on process scheduling and
other factors.)

This commit fixes the problem by explicitly ignoring the
exit status of the etcd command. Instead, check the output
of the command a little strictly.

Note: An alternative would be to disable pipefail in this script
as it's the only use of a pipe in this script. However, marun
prefers to keep it always enabled for any scripts.