Fix unjoin to use namespace identity as criteria for removal
Previously unjoin was comparing the context names of the host and
unjoining cluster, and removing the federation namespace in the
unjoining cluster if the names didn't match.  If the host context name
was not specified, though, a value of empty string (`""`) would be
used and the federation namespace would always be removed even for the
host cluster.  This PR switches to comparing the identities of the
federation namespace in the host and unjoining clusters and only
removing the namespace if the 2 don't match.