Fix service account propagation
Previously service account propagation had been configured to use the
generation field for versioning, which was masking the need to retain
the secrets field when it was populated by the in-cluster service
account controller.  Switching to the resource version field for
versioning resulted in thrashing between the sync and service
account controllers.  The sync controller would continually clear the
'secrets' field, and the service account controller would continually
generate a new token secret and set it in the 'secrets' field.  This
change ensures that the sync controller retains the 'secrets' field of
service accounts to avoid this problem.