Implement generic informer
The generated clients need to be removed as a precondition to
implementing HA (due to a client-go dependency), but the generic
client intended to replace the generated clients does not support the
watch call necessary for informers.  This commit hacks together a
generic informer to work around that omission so that the generic
client is a viable replacement for the generated clients.