Switch to an annotation to indicate cascade deletion intention
Previously the sync controller was ensuring the presence of 2
finalizers on federated resources that it reconciled:

 - `federation.kubernetes.io/delete-from-underlying-clusters`
 - `orphan`

The first finalizer was intended to ensure that a federation
controller would have the opportunity to perform pre-delete operations
(such as removing managed resources from member clusters).  The orphan
finalizer is a kube-internal finalizer used by the garbage collector,
and was used by federation to indicate that deletion of managed
resources should not be performed.  This reuse of a kube finalizer
proved problematic, though, since there is a non-zero chance that the
garbage collector could remove the orphan finalizer before the
federation sync controller could respond to it.

This commit switches to adding a new
finalizer (`federation.k8s.io/sync-controller`) to ensure that the
sync controller has a chance to perform pre-deletion operations, and
relies on the presence of a federation-specific
annotation (`federation.k8s.io/orphan: true`) to indicate that managed
resources should be orphaned.