Refactor cluster field retention for reuse
Previously the methods to retain cluster fields were in the same file
as the sync controller and returned the desired object even though the
original was being changed in-place.  This change isolates the functions
in a separate file to simplify maintenance, no longer returns the
desired object, and adds a function doc explaining that the object is
modified in-place.