Refactor update of resources in member clusters
The existing federated updater required that cluster operations be
pre-computed.  This complicated the proposal to cleanly implement
support for limiting the federation informers to managed resources and
then adopt resources as necessary (which would involve converting an
add to an update).

This change refactors the sync controller's reconcile loop and
implements a new updater to allow cluster operations to be dispatched
individually.  It also ensures that reconciliation events are logged
on federated resources where possible.