kubefed2 disable optionally deletes federated CRD
Allow the user to optionally delete the federated CRD type. If the
federated type config does not exist, such as when the user deleted it
first without providing the flag to also delete the federated CRD type,
a temporary federated type config is generated in order to construct the
proper CRD name for deletion of the federated CRD type.