Always retain metadata.finalizers and metadata.annotations
Since finalizers and annotations will typically be managed by local
controllers, ensure that these fields cannot be set from the template
and are always retained from an existing resource. It will still be
possible to apply overrides to add or remove entries from these
collections.

As part of this change, it was necessary to perform field retention
before overriding to ensure that retained fields could be the subject
of override directives.