Adding documentation to solve a webhook timeout
In GKE, the kubefed-controller fails to start up because it cannot
create new kubefed objects. This is because the API server attempts
to communicate to the Pod directly rather than going through
the service. This is intended behaviour when the API server runs
with the `--enable-aggregator-routing=true`.

An additional firewall rule within the VPC network firewall rules must
be added. The GKE installation guide has been extended to inform the
users of this issue.