Support varying the apiVersion of target resources
Previously the sync controller used ResourceClient to make changes to
member clusters. This meant that the API endpoint configuration for a
target type was always sourced from an FTC.

This commit updates the sync controller to use GenericClient so that
API endpoint configuration for create and update calls can be provided
on a per-call basis. If values for `apiVersion` is provided via
template or override for a given federated resource, that value will
supersede the configuration sourced from an FTC.

This is intended to allow a given federated type to manage more than
one version across member clusters. For example, federated resource
foo could manage a resource in cluster1 at v1, and a resource in
cluster2 at v2.