kubefedctl: Set host cluster context to current context if empty
For any invocation of `kubefedctl` that does not specify a value for
`--host-cluster-context`, the current config context will be
used. However, other usage of the `HostClusterContext` variable
requires explicitly looking up the name of the current context to
avoid using an empty string.

Previously the `join` command was fixed to lookup the current
context. This change ensures the same lookup is performed for all
`kubefedctl` commands.