Remove all metadata fields but labels from targets of kf federate
Previously `kubefedctl federate` removed a subset of metadata fields
from the resource targeted for conversion to a federated resource. It
wasn't removing annotations or finalizers despite those fields not
being supported for propagation. This change ensures that all fields
are removed except labels to ensure that fields set by the local
cluster and its controllers are not included in the template of the
new federated resource.