Refactor `kubefedctl join` for reuse by simulated scale testing
Previously, `kubefedctl join` assumed internally that the kubefed
namespace in the joining cluster should be the same as the kubefed
namespace in the host cluster. This refactor allows the kubefed system
namespace in the member cluster to vary in support of simulated scale
testing. This is not a user-visible change and is not intended for
external consumption.