Minimize propagation status updates
Each status update to a federated resource results in a subsequent
reconcile. Previously this was creating an endless cycle of
reconciliation even when there were no propagation changes since the
lastProbeTime timestamp was always changed even if nothing else
did. This induced unnecessary overhead in the controller manager, kube
apiserver, and etcd.

Note that Kubernetes had a similar problem with lastProbeTime and the
solution was to limit or stop updates to lastProbeTime:

https://github.com/kubernetes/kubernetes/issues/14393

This commit limits propagation status updates to when the status has
actually changed - either due to a propagation error or changes being
propagated to member clusters. This suggested replacing lastProbeTime
with lastUpdateTime.