Update APIEndpoint validation for go1.12.8 CVE fix
The url.Parse API now correctly reports an invalid port as part of the
Host in an APIEndpoint URL. This fix updates the error message reported
as we now expect to identify this error via the Parse API instead of
explicitly in the Atoi conversion. This CVE-2019-14809 was backported to
go1.12.8.