FIX: namespaced condition for api resource
The current code checks if the api resource is namespaced or not by a faulty condition,

	`apiResource.Namespaced = targetResource.GetNamespace() == ""`

If the input kubernetes api-resource has a namespace, the above condition makes the federatedapi-resource not namespaced, due to which the namespaces are not populated in the federatedapi-resource.

Changing the above condition to

	`apiResource.Namespaced = targetResource.GetNamespace() != ""`

makes the federated-api-resource as namespaced and sets the namespace as well.

Read more on [issue#1181](https://github.com/kubernetes-sigs/kubefed/issues/1181)