chore: make cluster creation work with kind v0.7.0
The command 'kind get kubeconfig-path' has been deprecated in kind
v0.6.0 and removed in v0.7.0 so the cluster creation script doesn't
work with the latter version. To account for this fact and other
changes (such as kind now respecting `KUBECONFIG` and falling back to
the same defaults that `kubectl` uses), I reworked some parts of the
script.

pre-commit now uses the kind imave v1.17.5 for tests.

Note: This change makes the scripts fail to work with
      kind version < v0.7.0.