feat: migrate off of openshift/generic-admission-server
That library os not actively maintained, anymore, and (worse) blocks
updating k8s dependencies to v0.18 and beyond because of breaking
changes to client-go (context.Context pass to most of the public
functions now).

Therefore, I refactored the webhook to rely on controller-runtime.
This also simplifies much of the webhook's code.