chore: webhook functions now return AdmissionResponse
Having these functions modify the AdmissionResponse passed in as
parameter makes it hard to reason about the code.

controller.webhook.util.Unmarshal was removed for the sake of not
having to return nil in case of success. This adds a bit more code to
the calling functions but at the same time helps bring across the
intended behaviour.