fix: prefer core resources when enabling a type
When a resource type such as `pods` is `services` is passed to
`kubefedctl enable` and there's a match for this name in the `core`
group as well as another group, there was no way to tell `kubefedctl`
that the `core` resource should be enabled.

Now, whenever an API resource from the `core` group matches the
desired type to be enabled, this resource type is preferred.