fix: value of retainReplicas isn't always respected
When creating a `FederatedDeployment` with `retainReplicas` set to
`true`, the `spec.replicas` value of the target Deployment isn't
touched by kubefed. However, when changing that value on the
`FederatedDeployment` back to `false`, the
`spec.template.spec.replicas` value isn't propagated to the target
resource.

Therefore, the `util.ObjectNeedsUpdate` func now compares the `spec`
field of federated resources so that any changes in it are actually
taken into account when deciding whether to update the target resource
or not.

refs #1278