fix: ignore non-targeted clusters during deletion
During the deletion of resources, the controller errors if any cluster
is marked as non-ready. That leads to objects being stuck during
deletion while these clusters are present, even if the resource is not
deployed in such clusters.

This commit makes the reconciliation for deletion to compute the
placement of the federated resources and ignore other clusters. This
way, the deletion will fail just if the non-ready clusters are the
clusters where the object is federated.