Add not-ready clusters E2E test
The e2e tests do not consider any unhealthy cluster. This commit adds a
test case where the cluster federates itself twice and makes one the
virtual federation not-ready by changing the API endpoint to an invalid
address. The ready cluster is labelled and a complete CRUD test
guarantees that the not-ready cluster does not impact operations if the
the cluster is not targeted.