build: Use KinD for syncing helm chart resources
Previously this was using kubebuilder provided API server and etcd, but these
were very out of date (k8s v1.16), leading to potentially out of date resources
and difficulties upgrading to later APIs that were not available in v1.16 (e.g.
`networking.k8s.io/v1`).

This commit fixes this by using KinD as the cluster to discover resources to
sync to the Helm chart.