build: Use setup-envtest to download later version of envtest
`controller-runtime` provides a convenient tool to download various versions
of k8s. This commit switches to using that to ensure we use a single version
for testing in unit tests and in e2e tests, namely v1.21.x right now so to
keep the version stable while just changing the installation mechanism.